+++ includes/custom-header/class-intro.php
@@ -61,13 +61,13 @@ class Reykjavik_Intro {
 
 						add_action( 'wp_enqueue_scripts', __CLASS__ . '::special_image', 120 );
 
-						add_filter( 'customize_partial_render_' . 'custom_header', __CLASS__ . '::special_image_partial_refresh' );
-
 					// Filters
 
 						add_filter( 'wmhook_reykjavik_intro_disable', __CLASS__ . '::disable', 5 );
 
 						add_filter( 'theme_mod_header_image', __CLASS__ . '::image', 15 ); // Has to be priority 15 for correct customizer previews.
+
+						add_filter( 'customize_partial_render_' . 'custom_header', __CLASS__ . '::special_image_partial_refresh' );
 
 		} // /__construct