+++ templates/parts/admin/welcome-demo.php
@@ -4,6 +4,8 @@
  *
  * Demo content installation.
  *
+ * @todo  OCDI plugin compatibility once you create demo XML.
+ *
  * @package    Reykjavik
  * @copyright  WebMan Design, Oliver Juhas
  *
@@ -30,7 +32,8 @@
 		<a href="https://www.webmandesign.eu/manual/reykjavik/#demo-content"><?php esc_html_e( 'Read the instructions &raquo;', 'reykjavik' ); ?></a>
 	</p>
 
-	<?php if ( ! class_exists( 'OCDI_Plugin' ) ) : ?>
+	<?php // if ( ! class_exists( 'OCDI_Plugin' ) ) : ?>
+	<?php if ( false ) : ?>
 
 		<a href="<?php echo esc_url( admin_url( 'themes.php?page=tgmpa-install-plugins' ) ); ?>" class="button button-hero"><strong><?php esc_html_e( 'Install and run "One Click Demo Import" plugin', 'reykjavik' ); ?></strong></a>